+++ src/utils/boxer2.zig
@@ -16,22 +16,22 @@ const VecA  = def.VecA;
 
 // ================================ UTIL FUNCTIONS ================================
 
-pub inline fn isLeftOf(  xVal : f32, thresholdX : f32 ) bool { return xVal < thresholdX; }
-pub inline fn isRightOf( xVal : f32, thresholdX : f32 ) bool { return xVal > thresholdX; }
-pub inline fn isAbove(   yVal : f32, thresholdY : f32 ) bool { return yVal < thresholdY; } // NOTE : Y axis is inverted in raylib rendering
-pub inline fn isBelow(   yVal : f32, thresholdY : f32 ) bool { return yVal > thresholdY; } // NOTE : Y axis is inverted in raylib rendering
+pub inline fn isLeftOf(  xVal : f64, thresholdX : f64 ) bool { return xVal < thresholdX; }
+pub inline fn isRightOf( xVal : f64, thresholdX : f64 ) bool { return xVal > thresholdX; }
+pub inline fn isAbove(   yVal : f64, thresholdY : f64 ) bool { return yVal < thresholdY; } // NOTE : Y axis is inverted in raylib rendering
+pub inline fn isBelow(   yVal : f64, thresholdY : f64 ) bool { return yVal > thresholdY; } // NOTE : Y axis is inverted in raylib rendering
 
-pub inline fn getCenterXFromLeftX(   leftX   : f32, scale : Vec2 ) f32 { return leftX   + scale.x; }
-pub inline fn getCenterXFromRightX(  rightX  : f32, scale : Vec2 ) f32 { return rightX  - scale.x; }
-pub inline fn getCenterYFromTopY(    topY    : f32, scale : Vec2 ) f32 { return topY    + scale.y; }
-pub inline fn getCenterYFromBottomY( bottomY : f32, scale : Vec2 ) f32 { return bottomY - scale.y; }
+pub inline fn getCenterXFromLeftX(   leftX   : f64, scale : Vec2 ) f64 { return leftX   + scale.x; }
+pub inline fn getCenterXFromRightX(  rightX  : f64, scale : Vec2 ) f64 { return rightX  - scale.x; }
+pub inline fn getCenterYFromTopY(    topY    : f64, scale : Vec2 ) f64 { return topY    + scale.y; }
+pub inline fn getCenterYFromBottomY( bottomY : f64, scale : Vec2 ) f64 { return bottomY - scale.y; }
 
 pub inline fn getCenterFromTopLeft(     topLeftPos     : Vec2, scale : Vec2 ) Vec2 { return Vec2{ .x = getCenterXFromLeftX(  topLeftPos.x,     scale ), .y = getCenterYFromTopY(    topLeftPos.y,     scale ) }; }
 pub inline fn getCenterFromTopRight(    topRightPos    : Vec2, scale : Vec2 ) Vec2 { return Vec2{ .x = getCenterXFromRightX( topRightPos.x,    scale ), .y = getCenterYFromTopY(    topRightPos.y,    scale ) }; }
 pub inline fn getCenterFromBottomLeft(  bottomLeftPos  : Vec2, scale : Vec2 ) Vec2 { return Vec2{ .x = getCenterXFromLeftX(  bottomLeftPos.x,  scale ), .y = getCenterYFromBottomY( bottomLeftPos.y,  scale ) }; }
 pub inline fn getCenterFromBottomRight( bottomRightPos : Vec2, scale : Vec2 ) Vec2 { return Vec2{ .x = getCenterXFromRightX( bottomRightPos.x, scale ), .y = getCenterYFromBottomY( bottomRightPos.y, scale ) }; }
 
-inline fn checkMinMax( minVal : f32, maxVal : f32 ) bool
+inline fn checkMinMax( minVal : f64, maxVal : f64 ) bool
 {
   if( minVal > maxVal )
   {
@@ -50,7 +50,7 @@ inline fn checkMinMax2( pMin : Vec2, pMax : Vec2 ) bool
   return false;
 }
 
-inline fn checkClampRange( minVal : f32, maxVal : f32, size : f32 ) bool
+inline fn checkClampRange( minVal : f64, maxVal : f64, size : f64 ) bool
 {
   if( maxVal - minVal < size )
   {
@@ -134,10 +134,10 @@ pub const Box2 = struct
     const sideStepAngle = Angle.newRad( def.TAU / @as( f32, @floatFromInt( sides )));
     const rP0 = Vec2.new( radii.x, 0.0 ).rot( a );
 
-    var maxX : f32 = rP0.x;
-    var minX : f32 = rP0.x;
-    var maxY : f32 = rP0.y;
-    var minY : f32 = rP0.y;
+    var maxX : f64 = rP0.x;
+    var minX : f64 = rP0.x;
+    var maxY : f64 = rP0.y;
+    var minY : f64 = rP0.y;
 
 
     if( sides > 2 ) // 1 == radius line, 2 == diametre line
@@ -176,20 +176,20 @@ pub const Box2 = struct
 
   // ================ ACCESSORS & MUTATORS ================
 
-  pub inline fn getLeftX(   self : *const Box2 ) f32 { return self.center.x - self.scale.x; }
-  pub inline fn getRightX(  self : *const Box2 ) f32 { return self.center.x + self.scale.x; }
-  pub inline fn getTopY(    self : *const Box2 ) f32 { return self.center.y - self.scale.y; }
-  pub inline fn getBottomY( self : *const Box2 ) f32 { return self.center.y + self.scale.y; }
+  pub inline fn getLeftX(   self : *const Box2 ) f64 { return self.center.x - self.scale.x; }
+  pub inline fn getRightX(  self : *const Box2 ) f64 { return self.center.x + self.scale.x; }
+  pub inline fn getTopY(    self : *const Box2 ) f64 { return self.center.y - self.scale.y; }
+  pub inline fn getBottomY( self : *const Box2 ) f64 { return self.center.y + self.scale.y; }
 
   pub inline fn getTopLeft(     self : *const Box2 ) Vec2 { return Vec2{ .x = self.getLeftX(),  .y = self.getTopY()    }; }
   pub inline fn getTopRight(    self : *const Box2 ) Vec2 { return Vec2{ .x = self.getRightX(), .y = self.getTopY()    }; }
   pub inline fn getBottomLeft(  self : *const Box2 ) Vec2 { return Vec2{ .x = self.getLeftX(),  .y = self.getBottomY() }; }
   pub inline fn getBottomRight( self : *const Box2 ) Vec2 { return Vec2{ .x = self.getRightX(), .y = self.getBottomY() }; }
 
-  pub inline fn setLeftX(   self : *Box2, leftX   : f32 ) void { self.center.x = leftX   + self.scale.x ; }
-  pub inline fn setRightX(  self : *Box2, rightX  : f32 ) void { self.center.x = rightX  - self.scale.x ; }
-  pub inline fn setTopY(    self : *Box2, topY    : f32 ) void { self.center.y = topY    + self.scale.y ; }
-  pub inline fn setBottomY( self : *Box2, bottomY : f32 ) void { self.center.y = bottomY - self.scale.y ; }
+  pub inline fn setLeftX(   self : *Box2, leftX   : f64 ) void { self.center.x = leftX   + self.scale.x ; }
+  pub inline fn setRightX(  self : *Box2, rightX  : f64 ) void { self.center.x = rightX  - self.scale.x ; }
+  pub inline fn setTopY(    self : *Box2, topY    : f64 ) void { self.center.y = topY    + self.scale.y ; }
+  pub inline fn setBottomY( self : *Box2, bottomY : f64 ) void { self.center.y = bottomY - self.scale.y ; }
 
   pub inline fn setTopLeft(     self : *Box2, topLeftPos     : Vec2 ) void { self.setLeftX(  topLeftPos.x     ); self.setTopY(    topLeftPos.y     ); }
   pub inline fn setTopRight(    self : *Box2, topRightPos    : Vec2 ) void { self.setRightX( topRightPos.x    ); self.setTopY(    topRightPos.y    ); }
@@ -203,23 +203,23 @@ pub const Box2 = struct
 
   // ======== PERMISSIVE RANGE CHECKERS ======== ( can be partially outside the range )
 
-  pub inline fn goesLeftOfX(  self : *const Box2, thresholdX : f32 ) bool { return isLeftOf(  self.getLeftX(),   thresholdX ); }
-  pub inline fn goesRightOfX( self : *const Box2, thresholdX : f32 ) bool { return isRightOf( self.getRightX(),  thresholdX ); }
-  pub inline fn goesAboveY(   self : *const Box2, thresholdY : f32 ) bool { return isAbove(   self.getTopY(),    thresholdY ); }
-  pub inline fn goesBelowY(   self : *const Box2, thresholdY : f32 ) bool { return isBelow(   self.getBottomY(), thresholdY ); }
+  pub inline fn goesLeftOfX(  self : *const Box2, thresholdX : f64 ) bool { return isLeftOf(  self.getLeftX(),   thresholdX ); }
+  pub inline fn goesRightOfX( self : *const Box2, thresholdX : f64 ) bool { return isRightOf( self.getRightX(),  thresholdX ); }
+  pub inline fn goesAboveY(   self : *const Box2, thresholdY : f64 ) bool { return isAbove(   self.getTopY(),    thresholdY ); }
+  pub inline fn goesBelowY(   self : *const Box2, thresholdY : f64 ) bool { return isBelow(   self.getBottomY(), thresholdY ); }
 
-  pub inline fn isOnX(     self : *const Box2, xVal  : f32  ) bool { if( self.isLeftOfX( xVal ) or self.isRightOfX( xVal )){ return false; } return true; }
-  pub inline fn isOnY(     self : *const Box2, yVal  : f32  ) bool { if( self.isAboveY(  yVal ) or self.isBelowY(   yVal )){ return false; } return true; }
+  pub inline fn isOnX(     self : *const Box2, xVal  : f64  ) bool { if( self.isLeftOfX( xVal ) or self.isRightOfX( xVal )){ return false; } return true; }
+  pub inline fn isOnY(     self : *const Box2, yVal  : f64  ) bool { if( self.isAboveY(  yVal ) or self.isBelowY(   yVal )){ return false; } return true; }
   pub inline fn isOnPoint( self : *const Box2, p     : Vec2 ) bool { if( !self.isOnX(    p.x  ) or !self.isOnY(     p.y  )){ return false; } return true; }
 
-  pub fn isOnXRange( self : *const Box2, xMin : f32, xMax : f32 ) bool
+  pub fn isOnXRange( self : *const Box2, xMin : f64, xMax : f64 ) bool
   {
     if( checkMinMax(  xMin, xMax )){ return false; }
     if( self.isLeftOfX(     xMin )){ return false; }
     if( self.isRightOfX(    xMax )){ return false; }
     return true;
   }
-  pub fn isOnYRange( self : *const Box2, minY : f32, maxY : f32 ) bool
+  pub fn isOnYRange( self : *const Box2, minY : f64, maxY : f64 ) bool
   {
     if( checkMinMax( minY, maxY )){ return false; }
     if( self.isAboveY(     minY )){ return false; }
@@ -236,12 +236,12 @@ pub const Box2 = struct
 
   // ======== RESTRICTIVE RANGE CHECKERS ======== ( must be entirely inside the range )
 
-  pub inline fn isLeftOfX(  self : *const Box2, thresholdX : f32 ) bool { return isLeftOf(  self.getRightX(),  thresholdX ); }
-  pub inline fn isRightOfX( self : *const Box2, thresholdX : f32 ) bool { return isRightOf( self.getLeftX(),   thresholdX ); }
-  pub inline fn isAboveY(   self : *const Box2, thresholdY : f32 ) bool { return isAbove(   self.getBottomY(), thresholdY ); }
-  pub inline fn isBelowY(   self : *const Box2, thresholdY : f32 ) bool { return isBelow(   self.getTopY(),    thresholdY ); }
+  pub inline fn isLeftOfX(  self : *const Box2, thresholdX : f64 ) bool { return isLeftOf(  self.getRightX(),  thresholdX ); }
+  pub inline fn isRightOfX( self : *const Box2, thresholdX : f64 ) bool { return isRightOf( self.getLeftX(),   thresholdX ); }
+  pub inline fn isAboveY(   self : *const Box2, thresholdY : f64 ) bool { return isAbove(   self.getBottomY(), thresholdY ); }
+  pub inline fn isBelowY(   self : *const Box2, thresholdY : f64 ) bool { return isBelow(   self.getTopY(),    thresholdY ); }
 
-  pub fn isInXRange( self : *const Box2, xMin : f32, xMax : f32 ) bool
+  pub fn isInXRange( self : *const Box2, xMin : f64, xMax : f64 ) bool
   {
     if( checkMinMax(       xMin, xMax )){                   return false; }
     if( checkClampRange(   xMin, xMax, self.scale.x * 2 )){ return false; }
@@ -249,7 +249,7 @@ pub const Box2 = struct
     if( self.goesRightOfX( xMax )){                         return false; }
     return true;
   }
-  pub fn isInYRange( self : *const Box2, minY : f32, maxY : f32 ) bool
+  pub fn isInYRange( self : *const Box2, minY : f64, maxY : f64 ) bool
   {
     if( checkMinMax(      minY, maxY )){                   return false; }
     if( checkClampRange(  minY, maxY, self.scale.y * 2 )){ return false; }
@@ -269,17 +269,17 @@ pub const Box2 = struct
 
   // ======== PERMISSIVE CLAMPERS ======== ( can be partially outside the range )
 
-  pub inline fn clampOnLeftX(   self : *Box2, thresholdX : f32 ) void { if( self.isLeftOfX(  thresholdX )){ self.setRightX(  thresholdX ); }}
-  pub inline fn clampOnRightX(  self : *Box2, thresholdX : f32 ) void { if( self.isRightOfX( thresholdX )){ self.setLeftX(   thresholdX ); }}
-  pub inline fn clampOnTopY(    self : *Box2, thresholdY : f32 ) void { if( self.isAboveY(   thresholdY )){ self.setBottomY( thresholdY ); }}
-  pub inline fn clampOnBottomY( self : *Box2, thresholdY : f32 ) void { if( self.isBelowY(   thresholdY )){ self.setTopY(    thresholdY ); }}
+  pub inline fn clampOnLeftX(   self : *Box2, thresholdX : f64 ) void { if( self.isLeftOfX(  thresholdX )){ self.setRightX(  thresholdX ); }}
+  pub inline fn clampOnRightX(  self : *Box2, thresholdX : f64 ) void { if( self.isRightOfX( thresholdX )){ self.setLeftX(   thresholdX ); }}
+  pub inline fn clampOnTopY(    self : *Box2, thresholdY : f64 ) void { if( self.isAboveY(   thresholdY )){ self.setBottomY( thresholdY ); }}
+  pub inline fn clampOnBottomY( self : *Box2, thresholdY : f64 ) void { if( self.isBelowY(   thresholdY )){ self.setTopY(    thresholdY ); }}
 
-  pub fn clampOnX( self : *Box2, xVal : f32 ) void
+  pub fn clampOnX( self : *Box2, xVal : f64 ) void
   {
     if( self.isLeftOfX(  xVal )){ self.setRightX( xVal ); }
     if( self.isRightOfX( xVal )){ self.setLeftX(  xVal ); }
   }
-  pub fn clampOnY( self : *Box2, yVal : f32 ) void
+  pub fn clampOnY( self : *Box2, yVal : f64 ) void
   {
     if( self.isAboveY(   yVal )){ self.setBottomY( yVal ); }
     if( self.isBelowY(   yVal )){ self.setTopY(    yVal ); }
@@ -290,13 +290,13 @@ pub const Box2 = struct
     self.clampOnY( p.y );
   }
 
-  pub fn clampOnXRange( self : *Box2, xMin : f32, xMax : f32 ) void
+  pub fn clampOnXRange( self : *Box2, xMin : f64, xMax : f64 ) void
   {
     if( checkMinMax( xMin, xMax )){ return; }
     self.clampOnLeftX(  xMin );
     self.clampOnRightX( xMax );
   }
-  pub fn clampOnYRange( self : *Box2, minY : f32, maxY : f32 ) void
+  pub fn clampOnYRange( self : *Box2, minY : f64, maxY : f64 ) void
   {
     if( checkMinMax( minY, maxY )){ return; }
     self.clampOnTopY(    minY );
@@ -309,7 +309,7 @@ pub const Box2 = struct
     self.clampOnYRange( pMin.y, pMax.y );
   }
 
-  pub fn clampNotInXRange( self : *Box2, xMin : f32, xMax : f32 ) void
+  pub fn clampNotInXRange( self : *Box2, xMin : f64, xMax : f64 ) void
   {
     if( checkMinMax( xMin, xMax )){ return; }
     const leftOverlap  = @abs( self.getRightX() - xMin );
@@ -318,7 +318,7 @@ pub const Box2 = struct
     if( leftOverlap < rightOverlap ){ self.clampOnLeftX(  xMin ); }
     else                            { self.clampOnRightX( xMax ); }
   }
-  pub fn clampNotInYRange( self : *Box2, minY : f32, maxY : f32 ) void
+  pub fn clampNotInYRange( self : *Box2, minY : f64, maxY : f64 ) void
   {
     if( checkMinMax( minY, maxY )){ return; }
     const topOverlap    = @abs( self.getBottomY() - minY );
@@ -337,17 +337,17 @@ pub const Box2 = struct
 
   // ======== RESTRICTIVE CLAMPERS ======== ( must be entirely inside the range )
 
-  pub inline fn clampInLeftX(   self : *Box2, thresholdX : f32 ) void { if( self.goesLeftOfX(  thresholdX )){ self.setLeftX(   thresholdX ); }}
-  pub inline fn clampInRightX(  self : *Box2, thresholdX : f32 ) void { if( self.goesRightOfX( thresholdX )){ self.setRightX(  thresholdX ); }}
-  pub inline fn clampInTopY(    self : *Box2, thresholdY : f32 ) void { if( self.goesAboveY(   thresholdY )){ self.setTopY(    thresholdY ); }}
-  pub inline fn clampInBottomY( self : *Box2, thresholdY : f32 ) void { if( self.goesBelowY(   thresholdY )){ self.setBottomY( thresholdY ); }}
+  pub inline fn clampInLeftX(   self : *Box2, thresholdX : f64 ) void { if( self.goesLeftOfX(  thresholdX )){ self.setLeftX(   thresholdX ); }}
+  pub inline fn clampInRightX(  self : *Box2, thresholdX : f64 ) void { if( self.goesRightOfX( thresholdX )){ self.setRightX(  thresholdX ); }}
+  pub inline fn clampInTopY(    self : *Box2, thresholdY : f64 ) void { if( self.goesAboveY(   thresholdY )){ self.setTopY(    thresholdY ); }}
+  pub inline fn clampInBottomY( self : *Box2, thresholdY : f64 ) void { if( self.goesBelowY(   thresholdY )){ self.setBottomY( thresholdY ); }}
 
-  pub fn clampNotOnX( self : *Box2, xVal : f32 ) void
+  pub fn clampNotOnX( self : *Box2, xVal : f64 ) void
   {
     if( self.center.x < xVal ){ self.clampInLeftX(  xVal ); }
     else                      { self.clampInRightX( xVal ); }
   }
-  pub fn clampNotOnY( self : *Box2, yVal : f32 ) void
+  pub fn clampNotOnY( self : *Box2, yVal : f64 ) void
   {
     if( self.center.y < yVal ){ self.clampInTopY(    yVal ); }
     else                      { self.clampInBottomY( yVal ); }
@@ -358,14 +358,14 @@ pub const Box2 = struct
     self.clampNotOnY( p.y );
   }
 
-  pub fn clampInXRange( self : *Box2, xMin : f32, xMax : f32 ) void
+  pub fn clampInXRange( self : *Box2, xMin : f64, xMax : f64 ) void
   {
     if( checkMinMax(      xMin, xMax )){ return; }
     if( checkClampRange(  xMin, xMax, self.scale.x * 2 )){ self.center.x = ( xMin + xMax ) * 0.5; return; }
     self.clampInLeftX(  xMin );
     self.clampInRightX( xMax );
   }
-  pub fn clampInYRange( self : *Box2, minY : f32, maxY : f32 ) void
+  pub fn clampInYRange( self : *Box2, minY : f64, maxY : f64 ) void
   {
     if( checkMinMax(     minY, maxY )){ return; }
     if( checkClampRange( minY, maxY, self.scale.y * 2 )){ self.center.y = ( minY + maxY ) * 0.5; return; }
@@ -379,7 +379,7 @@ pub const Box2 = struct
     self.clampInYRange( pMin.y, pMax.y );
   }
 
-  pub fn clampNotOnXRange( self : *Box2, xMin : f32, xMax : f32 ) void
+  pub fn clampNotOnXRange( self : *Box2, xMin : f64, xMax : f64 ) void
   {
     if( checkMinMax( xMin, xMax )){ return; }
     const leftOverlap  = @abs( self.getRightX() - xMin );
@@ -388,7 +388,7 @@ pub const Box2 = struct
     if( leftOverlap < rightOverlap ){ self.clampInLeftX(  xMin ); }
     else                            { self.clampInRightX( xMax ); }
   }
-  pub fn clampNotOnYRange( self : *Box2, minY : f32, maxY : f32 ) void
+  pub fn clampNotOnYRange( self : *Box2, minY : f64, maxY : f64 ) void
   {
     if( checkMinMax( minY, maxY )){ return; }
     const topOverlap    = @abs( self.getBottomY() - minY );
